Proposed update to default className
I propose that there is a default theme selected upon start. This makes it just a little bit easier when starting using vex for a few people. In my case I use brunch to bake vex into all of my stuff. I would love to not have a setup phase for vex if I am content with the plain vex-theme-os.